P1

+++ P1/src/ee/taltech/algoritmid/fibonacci/P1.java
@@ -15,11 +15,11 @@ public class P1 {
             return "1";
         }
 
-        int a = 0, b = 1; // initialiseerib nr
+        long a = 0, b = 1; // initialiseerib nr
 
         for (int i = 2; i <= n; i++) {  // i aka counter phm,2-st sest essal kahel vastus juba,
             // suurenda iga iteratsioon 1 vorra
-            int newB = a + b;
+            long newB = a + b;
             a = b;
             b = newB;
         }
